make escape sealed
